Update to correctly split dataObject, controller and view + Added start to opening times
