+++ JakesRestaurant/Program.cs
@@ -5,6 +5,8 @@ namespace JakesRestaurant
 {
     class Program
     {
+        private static Authentication.doUser currentUser;
+        public static Authentication.doUser MyUser { get { return currentUser; } set { currentUser = value; } }
         static void Main(string[] args)
         {
             ctlInitialise initialise = new ctlInitialise();
