+++ JakesRestaurant/model/dataObjects/openingTimes/doOpeningTimes.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace JakesRestaurant.model.dataObjects.openingTimes
+{
+    class doOpeningTimes
+    {
+    }
+}
